Fix User validations tests
