+++ spec/rails_helper.rb
@@ -5,7 +5,7 @@ require File.expand_path('../config/environment', __dir__)
 # Prevent database truncation if the environment is production
 abort("The Rails environment is running in production mode!") if Rails.env.production?
 require 'rspec/rails'
-require 'instance_helper'
+require 'helpers/model_helper'
 require 'helpers/controller_helper'
 require 'support/factory_bot'
 # Add additional requires below this line. Rails is not loaded until this point!
@@ -65,7 +65,7 @@ RSpec.configure do |config|
   # arbitrary gems may also be filtered via:
   # config.filter_gems_from_backtrace("gem name")
   
-  config.include(InstanceHelper)
+  config.include(ModelHelper)
   config.include(ControllerHelper)
 
   config.before(:suite) do
